+++ testsuite/execution_impl/my_testsuite_configurator.go
@@ -1,0 +1,32 @@
+package execution_impl
+
+import (
+	"github.com/kurtosis-tech/kurtosis-onboarding-experience/testsuite/testsuite_impl"
+	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
+	"github.com/palantir/stacktrace"
+	"github.com/sirupsen/logrus"
+)
+
+type MyTestsuiteConfigurator struct {}
+
+func NewMyTestsuiteConfigurator() *MyTestsuiteConfigurator {
+	return &MyTestsuiteConfigurator{}
+}
+
+func (t MyTestsuiteConfigurator) SetLogLevel(logLevelStr string) error {
+	level, err := logrus.ParseLevel(logLevelStr)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred parsing loglevel string '%v'", logLevelStr)
+	}
+	logrus.SetLevel(level)
+	logrus.SetFormatter(&logrus.TextFormatter{
+		ForceColors:   true,
+		FullTimestamp: true,
+	})
+	return nil
+}
+
+func (t MyTestsuiteConfigurator) ParseParamsAndCreateSuite(paramsJsonStr string) (testsuite.TestSuite, error) {
+	suite := &testsuite_impl.MyTestsuite{}
+	return suite, nil
+}