+++ docs/changelog.md
@@ -1,4 +1,8 @@
 # TBD
+#Features
+* Created the Kurtosis project boilerplate which includes: CI configuration, build and run script and release script
+* Added testsuite execution configuration with a custom configurator `my_testsuite_configurator` and a custom testsuite `my_testsuite`
+* Added a basic ETH test `my_test` which test GETH client execution following the GETH documentation for [setting up a GETH client in DEV mode](https://geth.ethereum.org/docs/getting-started/dev-mode)
 
 #0.1.0
 * Init commit